#80 Search widgets and tables do not correspond to each other
Cleanup

SD_WLO-360
